+++ qgreenland/config/layers/Places/__settings__.py
@@ -2,7 +2,7 @@ from qgreenland.models.config.layer_group import LayerGroupSettings
 
 settings = LayerGroupSettings(
     order=[
-        ":cities",
+        ":towns",
         ":settlements",
         ":points_of_interest",
         ":roads",